Anyadidos extractores

+++ CentroSaludCore/appcore/models.py
@@ -4,14 +4,14 @@ from django.db import models
 
 class Provincia(models.Model):
     nombre = models.CharField(max_length=50)
-    codigo = models.PositiveSmallIntegerField(primary_key=True)
+    codigo = models.PositiveIntegerField(primary_key=True)
 
     def __str__(self):
         return self.nombre
 
 class Localidad(models.Model):
     nombre = models.CharField(max_length=50)
-    codigo = models.PositiveSmallIntegerField(primary_key=True)
+    codigo = models.PositiveIntegerField(primary_key=True)
     en_provincia = models.ForeignKey(Provincia, on_delete=models.CASCADE)
     
     def __str__(self):
